Refactor theme-preval
Turns out preval was silently not working because it doesn't parse TypeScript
