+++ package.json
@@ -84,6 +84,8 @@
     "@types/enzyme": "3.10.8",
     "@types/jest": "26.0.20",
     "@types/jest-axe": "3.5.1",
+    "@types/lodash.isempty": "4.4.6",
+    "@types/lodash.isobject": "3.0.6",
     "@typescript-eslint/eslint-plugin": "4.14.1",
     "@typescript-eslint/parser": "4.14.1",
     "@wojtekmaj/enzyme-adapter-react-17": "0.3.2",
@@ -98,6 +100,7 @@
     "chroma-js": "2.1.0",
     "concurrently": "5.3.0",
     "cross-env": "7.0.2",
+    "deepmerge": "4.2.2",
     "enzyme": "3.10.0",
     "eslint": "6.5.1",
     "eslint-plugin-github": "4.1.1",
@@ -110,6 +113,8 @@
     "jest-axe": "4.1.0",
     "jest-styled-components": "6.3.3",
     "jscodeshift": "0.6.4",
+    "lodash.isempty": "4.4.0",
+    "lodash.isobject": "3.0.2",
     "playroom": "0.15.1",
     "react": "^17.0.0",
     "react-dom": "^17.0.0",
